avoid modal being closed twice
