+++ src/components/Modal/index.jsx
@@ -1,21 +1,10 @@
 // @flow
-import _ from 'lodash'
-import { ReactElement } from 'tcomb-react'
 import $ from 'jquery'
-import autobind from 'autobind-decorator'
-import cn from 'classnames/bind'
-import Form from '../Form'
-import Icon from '../Icon'
-import Illustration from '../Illustration'
+import Content from './Content'
 import Portal from 'react-portal'
 import React from 'react'
-import RoundButton from '../Buttons/RoundButton'
-
-import styles from './index.scss'
 
 import type { Modal as ModalType } from './types'
-
-const cx = cn.bind(styles)
 
 export default class Modal extends React.Component {
   props: ModalType
@@ -38,151 +27,11 @@ export default class Modal extends React.Component {
     })
   }
 
-  @autobind
-  cancel () {
-    this.props.onClose()
-  }
-
-  @autobind
-  submit () {
-    const { onSubmit } = this.props
-
-    if (onSubmit) {
-      return onSubmit()
-    } else {
-      return Promise.resolve()
-    }
-  }
-
-  @autobind
-  onCancel (event: any) {
-    event.preventDefault()
-    this.cancel()
-  }
-
-  @autobind
-  onClickBox (event: any) {
-    event.stopPropagation()
-  }
-
-  @autobind
-  onClickOverlay () {
-    this.cancel()
-  }
-
-  getClassNames () {
-    const classes = [styles.modal]
-
-    // Only used for the styleguide
-    if (!_.isNil(this.props.fixed) && !this.props.fixed) {
-      classes.push([styles.noFixed])
-    }
-
-    return classes.join(' ')
-  }
-
-  @autobind
-  renderElement (element: ?ReactElement) {
-    if (!element) return null
-
-    return (
-      <div className={styles.element}>
-        {element}
-      </div>
-    )
-  }
-
-  renderIllustration () {
-    const { color, illustration, type } = this.props
-
-    if (!illustration) return null
-
-    if (typeof illustration !== 'string') {
-      return (
-        <div
-          className={styles.illustration}
-          style={{ backgroundColor: illustration.backgroundColor }}
-        >
-          {illustration.illustration}
-        </div>
-      )
-    }
-
-    return (
-      <div className={styles.illustration}>
-        <Illustration color={color} type={type} name={illustration} />
-      </div>
-    )
-  }
-
-  renderCancel () {
-    const { locked, negative } = this.props
-
-    if (locked) return null
-
-    return (
-      <div className={styles.close}>
-        <button type='button' onClick={this.onCancel}>
-          <Icon
-            type={negative ? 'negative' : 'primary'}
-            set='utility'
-            icon='close'
-            size='small'
-          />
-        </button>
-      </div>
-    )
-  }
-
-  renderBack () {
-    const { onBack } = this.props
-    if (!onBack) return null
-    return (
-      <div className={styles.back}>
-        <RoundButton
-          icon='back'
-          onClick={onBack}
-          set='utility'
-          type='negative'
-        />
-      </div>
-    )
-  }
-
-  renderHeaderContent () {
-    const { headerContent } = this.props
-    if (!headerContent) return null
-    return headerContent()
-  }
-
   renderContent () {
-    const { title, description, big, negative, children } = this.props
-
     return (
-      <div className={this.getClassNames()}>
-        <div className={styles.overlay} onClick={this.onClickOverlay}>
-          <div
-            className={cx('box', { big, negative })}
-            onClick={this.onClickBox}
-          >
-            <Form onSubmit={this.submit} className={styles.form}>
-              {this.renderCancel()}
-              {this.renderBack()}
-              <div className={styles.header}>
-                {this.renderIllustration()}
-                <h2 className={styles.title}>
-                  {title}
-                </h2>
-                <h3 className={styles.description}>
-                  {description}
-                </h3>
-                {this.renderHeaderContent()}
-              </div>
-              {React.Children.map(children, this.renderElement)}
-            </Form>
-          </div>
-        </div>
-      </div>
+      <Content {...this.props}>
+        {this.props.children}
+      </Content>
     )
   }
 
@@ -198,7 +47,7 @@ export default class Modal extends React.Component {
     }
 
     return (
-      <Portal closeOnEsc isOpened onClose={this.cancel}>
+      <Portal closeOnEsc isOpened onClose={this.props.onClose}>
         {this.renderContent()}
       </Portal>
     )
